requirements.txt
numpy==1.21.5
matplotlib==3.5.2
gradio==3.23.0
networkx==2.8.4
requests==2.28.1
beautifulsoup4==4.11.1
pillow==9.3.0